- When fields are empty/removed they shouldn't appear in the data. For example: check the palette field in emus in the "tree" pack. Please fix this once FOR ALL possible fields and not just for specific fields. fixed
- Field values MUST be stored in the correct field type.
For example, sometime duration or frame numbers are stored as a string instead of a numeric value.
Each field type was defined for you in the original spec. Check it out.
fixed

- All field names should be in lower case + underscore notation.
For example, patchedOn should be patched_on.
fixed

- Back layer shouldn't be optional. Make it mandatory.
The client assumes all emus have at least a back layer
(it needs to be able to stitch the extracted user layer to something)
fixed

- Rename the field name storing the name of the latest zipped package.
Use this field name: zipped_package_file_name
fixed

- The zip button is enabled not only as we defined.
We defined that the ZIP button for creating a new zip file
with the resources should only be enabled when a new zip file is required.
A new zip file is required ONLY when any of the emus layer resources changed/updated.
Don't enable the zip button if simple info about the pack or icons of the pack were updated.
fixed

- Add new pack button not working.
fixed

- In the UI: Icon Mask should be User Mask
The field got nothing to do with icons. It is the user mask layer used in rendering.
fixed

- In the UI: It is ok that the UI isn't designed, but fields should be readable to the content manager.
Some of the fields appear on the same line or are clustered together
fixed